python/requirements.txt
Cython>=0.19.2
numpy>=1.7.1
scipy>=0.17.0
scikit-image>=0.9.3
matplotlib>=1.3.1
ipython>=3.0.0
h5py>=2.2.0
leveldb>=0.191
networkx>=1.8.1
nose>=1.3.0
pandas>=0.12.0
python-dateutil>=2.6.0
protobuf>=2.5.0
python-gflags>=2.0
pyyaml>=3.10
Pillow>=5.1.0
six>=1.10.0
lxml>=4.2.2
opencv-python>=3.3.0

